change download

+++ app/src/main/java/com/example/wwez/Imooc_download/com/download/entry/ThreadInfo.java
@@ -48,7 +48,7 @@ public class ThreadInfo {
         return finished;
     }
 
-    public void setFinished(int finished) {
+    public void setFinished(long finished) {
         this.finished = finished;
     }
